% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexbin.R
\name{add_hexbin}
\alias{add_hexbin}
\title{Hexbin Layers for Leaflet Plots}
\usage{
add_hexbin(map, data = NULL, radius = 8.5, opacity = 0.5,
  duration = 500, lowEndColor = "white", highEndColor = "blue",
  uniformSize = FALSE, uniformColor = NULL)
}
\arguments{
\item{map}{The leaflet map object to apply the hexbin layer to. Makes this function compatible with the %>% operator}

\item{data}{data frame or tibble - alternate data to use for this hexbin instead of default map data}

\item{radius}{choose the base size for the hexagons}

\item{opacity}{decimal between 0.0 and 1.0 - choose the percent of opacity for the hexagons}

\item{lowEndColor}{choose the color for the smaller hexagons}

\item{highEndColor}{choose the color for the larger hexagons}

\item{uniformSize}{boolean for having uniformly sized hexagons or smaller hexagons for area containing fewer data points}

\item{uniformColor}{a color that overrides lowEndColor and highEndColor to make the color uniform across the hexagon sizes.}
}
\value{
map parameter, but with the hexbinLayer attached so that it can be used with the %>% pipe operator
}
\description{
Create hexbin layers for leaflet plots.
}
\note{
Do not use uniformColor and uniformSize together as it will not give any insights to the data
}
