% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addJSPlugin.R
\name{pluginFactory}
\alias{pluginFactory}
\title{Create plugin functions to implement Leaflet pure-JS libraries in R}
\usage{
pluginFactory(name = "JSPlugin", location, jsfilename, dependencies,
  stylesheet = NULL)
}
\description{
pluginFactory takes in a couple js files (code and dependencies)
and creates a plugin function that applies this plugin to a given map. The new
function also returns the map so it can be piped with the %>% operator like other
leaflet functions
}
