% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexbin.R
\name{add_hexbin}
\alias{add_hexbin}
\title{Hexbin Layers for Leaflet Plots}
\usage{
add_hexbin(map, data = NULL, radius = 8.5, opacity = 0.5,
  duration = 500, lowEndColor = "white", highEndColor = "blue")
}
\description{
Create hexbin layers for leaflet plots.
}
