% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addJSPlugin.R
\name{pluginFactory}
\alias{pluginFactory}
\title{generic plugin Factory for pure JS leaflet libraries}
\usage{
pluginFactory(name = "JSPlugin", location, jsfilename, dependencies,
  stylesheet = NULL)
}
\description{
Create plugin functions to implement pure JS
}
